co-existence with cordova-plugin-local-notification on iOS
Local notifications on iOS via cordova-plugin-local-notification are working again.
Click-callback for remote notifications not working right now.

Details:
UNUserNotificationCenter accepts a single delegate that handles the display of notifications. cordova-plugin-firebase now stores a reference to the existing delegate if any and proxies the delegate-calls (cordova-plugin-local-notification is already doing that in the other direction when applicable).