Prevent reconfiguration causing exception
We ran into a problem with a configuration using both the firebasex and firebase-analytics plugins. The latter was doing "[FIRApp configure]" during the "[self application:application..." super call at the top of this function, which caused an exception when FIRApp was configured again by this plugin. The exception bypassed the remainder of the code block and prevented our app from receiving push notification data.